requirements.txt
flask
flask-cors
flask-sqlalchemy
requests
beautifulsoup4
python-dotenv
google-generativeai
